+++ src/components/nodes/endEvent/endEvent.vue
@@ -24,7 +24,7 @@ import { endColor, endColorStroke } from '@/components/nodeColors';
 import CrownConfig from '@/components/crown/crownConfig/crownConfig';
 import highlightConfig from '@/mixins/highlightConfig';
 import defaultNames from '@/components/nodes/endEvent/defaultNames';
-import store from '@/store';
+import documentingIcons from '@/mixins/documentingIcons';
 
 
 export default {
@@ -44,7 +44,7 @@ export default {
     'planeElements',
     'isRendering',
   ],
-  mixins: [highlightConfig, portsConfig, hideLabelOnDrag],
+  mixins: [highlightConfig, portsConfig, hideLabelOnDrag, documentingIcons],
   data() {
     return {
       shape: null,
@@ -101,37 +101,7 @@ export default {
     this.shape.addTo(this.graph);
     this.shape.component = this;
 
-    const docElement = this.node?.definition?.documentation;
-    const doc = Array.isArray(docElement)
-      ? (docElement[0].text ?? '').trim()
-      : (docElement ?? '').trim();
-
-    const view = this.paper.findViewByModel(this.shape);
-    view.model.attr({
-      doccircle: {
-        display:'none',
-      },
-      doclabel: {
-        display: 'none',
-        style: 'text-anchor: middle; transform: translate(30px, -4px);',
-        text: null,
-      },
-    });
-
-    const interval = window.setInterval(() => {
-      if (view.$('circle').length > 0 && store.getters.isForDocumenting) {
-        view.model.attr({
-          doccircle: {
-            display:(doc ? 'block' : 'none'),
-          },
-          doclabel: {
-            display: 'none',
-          },
-        });
-        clearInterval(interval);
-      }
-    }, 200);
-
+    this.initDocumentingIcons({labelX: '30px', labelY: '-4px'});
   },
 };
 </script>
